Feature/migrate

+++ blog/views.py
@@ -1,5 +1,5 @@
-from flask import render_template, request, redirect, url_for, flash
-from flask.ext.login import login_user, login_required
+from flask import render_template, request, redirect, url_for, flash, abort
+from flask.ext.login import login_user, login_required, current_user, logout_user
 from werkzeug.security import check_password_hash
 
 from . import app
@@ -62,8 +62,10 @@ def add_entry_get():
 @login_required
 def add_entry_post():
     """ add entry to DB """
-    entry = Entry(title=request.form["title"],
-        content=request.form["content"]
+    entry = Entry(
+        title=request.form["title"],
+        content=request.form["content"],
+        author=current_user
     )
     session.add(entry)
     session.commit()
@@ -78,25 +80,26 @@ def single_entry_get(id):
             entry=entry
             )
     else:
-        return redirect(url_for("error_route"))
+        abort(404)
         
 @app.route('/entry/<int:id>/edit',methods=['GET'])
 @login_required
 def edit_entry(id):
     """ get form to edit single entry """
+    
     entry = get_entry(id)
-    if entry:
+    if entry and entry.author and entry.author.id == current_user.id:
         return render_template("edit_entry.html",entry=entry)
     else:
-        return redirect(url_for("error_route"))
+        abort(404)
     
 @app.route('/entry/<int:id>/edit',methods=['POST'])
 @login_required
 def update_entry(id):
     """ post route for editing entry """
     entry = get_entry(id)
-    if not entry:
-         return redirect(url_for("error_route"))
+    if not entry or not entry.author or entry.author.id != current_user.id:
+         abort(404)
          
     entry.title = request.form["title"]
     entry.content =request.form["content"]
@@ -108,14 +111,17 @@ def update_entry(id):
 def delete_entry(id):
     """ delete entry """
     entry = get_entry(id)
-    session.delete(entry)
-    session.commit()
-    return redirect(url_for("entries"))
-
-@app.route('/404')
-def error_route():
+    if entry.author and entry.author.id == current_user.id:
+        session.delete(entry)
+        session.commit()
+        return redirect(url_for("entries"))
+    else:
+        abort(404)
+        
+@app.errorhandler(404)
+def error_route(e):
     """ return 404 """
-    return render_template("404.html")
+    return render_template('404.html'), 404
     
 @app.route("/login",methods=["GET"])
 def login_get():
@@ -133,3 +139,9 @@ def login_post():
         return redirect(url_for("login_get"))
     login_user(user)
     return redirect(request.args.get('next') or url_for("entries"))
+
+@app.route('/logout')
+def logout():
+    """ logout current user """
+    logout_user()
+    return redirect(url_for('entries'))
